% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cud_color.R
\name{cud_color}
\alias{cud_color}
\title{Get CUD color code}
\usage{
cud_color(colname)
}
\arguments{
\item{colname}{The vector of color names to convert to hexadecimal color codes.}
}
\value{
vector of hexadecimal color codes (#rrggbb).
}
\description{
Get Color Universal Design or JIS Z 9103:2018 color code(s) from color name(s).
The color name(s) for JIS color should be specified with a "jis_" prefix.
}
\seealso{
Other color: 
\code{\link{jis_color}()}
}
\concept{color}
