% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jis_color.R
\name{jis_color}
\alias{jis_color}
\title{Get JIS Z 9103:2018 color code}
\usage{
jis_color(colname)
}
\arguments{
\item{colname}{The vector of color names to convert to hexadecimal color codes.}
}
\value{
vector of hexadecimal color codes (#rrggbb).
}
\description{
Get JIS Z 9103:2018 color code
}
\seealso{
Other color: 
\code{\link{cud_color}()}
}
\concept{color}
