% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jis_palette.R
\name{jis_pal}
\alias{jis_pal}
\title{create palette function using JIS Z 9103:2018 color palette}
\usage{
jis_pal(direction = 1, reorder = TRUE, warn = TRUE)
}
\arguments{
\item{direction}{Sets the order of colors. If -1, the order of colors is reversed.}

\item{reorder}{a logical (default is TRUE).
If TRUE, the order of colors is reorderd to maximize color difference between adjacent colors.
If FALSE, the order is along with the order described in reference documents.}

\item{warn}{a logical (default is TRUE) indicating whether puts warning message for invalid parameters.}
}
\value{
color palette function (closure)
}
\description{
This is a wrapper function calls `cud_pal()` with type = "jis".
}
\seealso{
Other palette: 
\code{\link{cud_pal}()}
}
\concept{palette}
