% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cud_color.R
\name{cud_color}
\alias{cud_color}
\title{Get CUD color code}
\usage{
cud_color(colname)
}
\arguments{
\item{colname}{The vector of color names to convert to hexadecimal color codes.}
}
\value{
vector of hexadecimal color codes (#rrggbb).
}
\description{
Get CUD color code
}
