% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cud_palette.R
\name{cud_pal}
\alias{cud_pal}
\title{create palette function using CUD model color palette}
\usage{
cud_pal(type = "accent", size = 6, pal = 1, direction = 1, warn = TRUE)
}
\arguments{
\item{type}{The name of color category (default is "accent").}

\item{size}{The number of colors in palette (default is 6).}

\item{pal}{The number of the palette (default is 1).}

\item{direction}{Sets the order of colors. If -1, the order of colors is reversed.}

\item{warn}{a logical (default is TRUE) indicating whether puts warning message for invalid parameters.}
}
\value{
color palette function (closure)
}
\description{
create palette function using CUD model color palette
}
